Moved common cart functionality from separate containers to a helper file
