+++ app/components/cart/CartContainer.jsx
@@ -3,16 +3,7 @@
 import Cart from './Cart';
 import { connect } from 'react-redux';
 import { setCart } from '../../reducers/cart';
-
-const getCartFromLocal = () => dispatch => {
-  const local = window.localStorage.getItem('cart');
-  if(local) {
-    const windowCart = JSON.parse(local);
-    if (Array.isArray(windowCart)) {
-      dispatch(setCart(windowCart));
-    }
-  }
-};
+import { getCartFromLocal } from './CartHelpers';
 
 const mapStateToProps = ({ cart }, { params }) => ({
   cart
