+++ app/components/cart/CartHelpers.js
@@ -1,0 +1,19 @@
+'use strict';
+
+export const getCartFromLocal = () => dispatch => {
+  const local = window.localStorage.getItem('cart');
+  if(local) {
+    const windowCart = JSON.parse(local);
+    if (Array.isArray(windowCart)) {
+      dispatch(setCart(windowCart));
+    }
+  }
+};
+
+export const addToCartThunk = product => dispatch => {
+  let local = window.localStorage.getItem('cart');
+  if (!local) { local = '[]'; }
+  const windowCart = JSON.parse(local);
+  window.localStorage.setItem('cart', JSON.stringify([...windowCart, product]));
+  dispatch(addToCart(product));
+}
